refactor: change structure @reduxjs/toolkit
